extracts BFF fetch mechanism into utility
