+++ src/app/routes/article/getInitialData/index.ts
@@ -1,36 +1,17 @@
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 import { Agent } from 'https';
-import pipe from 'ramda/src/pipe';
-import Url from 'url-parse';
 import getAdditionalPageData from '#app/routes/cpsAsset/utils/getAdditionalPageData';
 import getEnvironment from '#app/routes/utils/getEnvironment';
 import nodeLogger from '../../../lib/logger.node';
-import { getUrlPath } from '../../../lib/utilities/urlParser';
 import { BFF_FETCH_ERROR } from '../../../lib/logger.const';
 import fetchPageData from '../../utils/fetchPageData';
+import constructBffUrl from '../../utils/constructBffUrl';
+import handleError from '../../utils/handleError';
 import { Services, Variants } from '../../../models/types/global';
 import getOnwardsPageData from '../utils/getOnwardsData';
 import { advertisingAllowed, isSfv } from '../utils/paramChecks';
 
 const logger = nodeLogger(__filename);
-
-const removeAmp = (path: string) => path.split('.')[0];
-const getOptimoId = (path: string) => path.match(/(c[a-zA-Z0-9]{10}o)/)?.[1];
-const getCpsId = (path: string) => path;
-
-const getId = (pageType: string) =>
-  pipe(getUrlPath, removeAmp, pageType === 'article' ? getOptimoId : getCpsId);
-
-interface BFFError extends Error {
-  status: number;
-}
-
-const handleError = (message: string, status: number) => {
-  const error = new Error(message) as BFFError;
-  error.status = status;
-
-  return error;
-};
 
 type Props = {
   getAgent: () => Promise<Agent>;
@@ -52,32 +33,10 @@ export default async ({
     const isLocal = !env || env === 'local';
 
     const agent = !isLocal ? await getAgent() : null;
-    const id = getId(pageType)(pathname);
 
-    if (!id) throw handleError('Article ID is invalid', 500);
-
-    let fetchUrl = Url(process.env.BFF_PATH as string).set('query', {
-      id,
-      service,
-      ...(variant && {
-        variant,
-      }),
-      pageType,
-    });
-
+    const urlParams = { pathname, pageType, service, variant };
+    const fetchUrl = constructBffUrl(urlParams);
     const optHeaders = { 'ctx-service-env': env };
-
-    if (isLocal) {
-      if (pageType === 'article') {
-        fetchUrl = Url(
-          `/${service}/articles/${id}${variant ? `/${variant}` : ''}`,
-        );
-      }
-
-      if (pageType === 'cpsAsset') {
-        fetchUrl = Url(id);
-      }
-    }
 
     // @ts-ignore - Ignore fetchPageData argument types
     // eslint-disable-next-line prefer-const
