+++ src/app/routes/utils/handleError/index.ts
@@ -1,0 +1,12 @@
+interface BFFError extends Error {
+  status: number;
+}
+
+const handleError = (message: string, status: number) => {
+  const error = new Error(message) as BFFError;
+  error.status = status;
+
+  return error;
+};
+
+export default handleError;
